EvoNN-Contenders/pyproject.toml
[project]
name = "evonn-contenders"
version = "0.1.0"
description = "Fixed contender zoo for EvoNN benchmark packs"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.1.3",
    "numpy>=2.2.0",
    "openml>=0.14.2",
    "pandas>=2.2.3",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
    "rich>=13.9.0",
    "scikit-learn>=1.6.1",
    "typer>=0.15.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4"]

[project.scripts]
evonn-contenders = "evonn_contenders.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/evonn_contenders"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["evonn_contenders"]

[tool.coverage.report]
show_missing = true
skip_covered = true
